+++ Amplitude/AMPIdentify.h
@@ -12,6 +12,7 @@
 
 + (instancetype)identify;
 - (AMPIdentify*)add:(NSString*) property value:(NSObject*) value;
+- (AMPIdentify*)append:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)set:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)setOnce:(NSString*) property value:(NSObject*) value;
 - (AMPIdentify*)unset:(NSString*) property;